Cleanup state converters
