+++ src/modules/src/Eryph.Modules.VmHostAgent/Inventory/InventoryRequestedEventHandler.cs
@@ -77,7 +77,7 @@ internal class InventoryRequestedEventHandler(IBus bus, IPowershellEngine engine
 
     private bool IsInventorizable(TypedPsObject<VirtualMachineInfo> vmInfo)
     {
-        var operationalStatus = OperationalStatusConverter.Convert(vmInfo.Value.OperationalStatus);
+        var operationalStatus = VmStateUtils.Convert(vmInfo.Value.OperationalStatus);
         var state = vmInfo.Value.State;
         var isInventorizable = VmStateUtils.isInventorizable(state, operationalStatus);
