+++ src/modules/src/Eryph.Modules.VmHostAgent/Inventory/VirtualMachineChangedEventHandler.cs
@@ -71,7 +71,7 @@ internal class VirtualMachineChangedEventHandler(
 
     private bool IsInventorizable(TypedPsObject<VirtualMachineInfo> vmInfo)
     {
-        var operationalStatus = OperationalStatusConverter.Convert(vmInfo.Value.OperationalStatus);
+        var operationalStatus = VmStateUtils.Convert(vmInfo.Value.OperationalStatus);
         var state = vmInfo.Value.State;
         var isInventorizable = VmStateUtils.isInventorizable(state, operationalStatus);
